requirements.txt
Pillow==10.1.0
Requests==2.31.0
streamlit==1.28.2
streamlit_lottie==0.0.5
pybase64==1.3.1